Add cryptographic (SRTP) negotiation to SDP
Update SipAudioCall.java to do sdp negotiation of RTP/SAVP (if TLS is
selected as the transport) first and fall back to RTP/AVP later.  The
standard (RFC 4568) says that only one cipher can be used for the
transmission, so this is in effect a cipher agreement but that we must
have separate keys for sending and recieving which are sent by both
parties in SDP.  For this reason, we have to have access to both the
SDP we sent and the one we received.  This patch looks bigger because
it adds looping over the media type for the both RTP/AVP and RTP/SAVP
case.

Change-Id: I28eaf39199dea7a31497bace67085c7bf30c9792
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>