JNI: add SRTP processing using external cisco libsrtp2
Expand the nativeAdd method to take a cipher and the in and out keys.
These are negotiated in the session description exchange.  They should
either all be set or none of them (indicating no encryption).  Note
the session is negotiated and tested with asterisk which doesn't do
key lifetime or multiple master keys, so this is a very simple SRTP
setup.  Note that the encryption for SRTP includes authentication,
thus the packet expands on encryption and contracts on decryption,
which is why the length is adjusted coming out of
srtp_unprotect/protect.  We also need to allocate additional space in
the buffers to cope with this expansion.

Change-Id: Ic438756774a5f4eb2986a28a75f4066ceae111c1
Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>